feat: Complete VoiceStand Learning System v2.0
- Docker-based learning system with local-only security (127.0.0.1:7890)
- PostgreSQL persistent storage with pgvector for audio features
- Dark-themed real-time dashboard with model indicators
- C++ API integration with non-blocking HTTP client
- UK English specialization targeting 94-99% accuracy
- Ensemble processing with adaptive model weights
- Training interface for submitting learning examples
- Fixed compilation issues and type conflicts
- Added comprehensive documentation and setup scripts

+++ src/core/personal_voice_commands.cpp
@@ -701,7 +701,7 @@ void PersonalVoiceCommands::cleanup_caches() {
 
     while (it != dtw_cache_.end()) {
         auto age = std::chrono::duration_cast<std::chrono::milliseconds>(now - it->second.timestamp).count();
-        if (age > config_.pattern_cache_ms) {
+        if (age > static_cast<long>(config_.pattern_cache_ms)) {
             it = dtw_cache_.erase(it);
         } else {
             ++it;
